refactor(sdk): Discusser/Executor가 callLLMWithFallback을 사용하도록 마이그레이션
오케스트레이션 일반화 마이그레이션 B — SDK 진입점에 자동 폴백 적용.
import 별칭 한 줄 변경으로 본문 코드 변경 0. 시그니처/응답 호환.

- 첫 모델이 429/5xx 또는 네트워크 에러로 실패 시 ModelSelector.selectFallback
  체인 따라 다음 모델 자동 시도 (opus → sonnet → haiku)
- 401/403 등 권한 에러는 즉시 throw (기존과 동일)
- consultation/review-conversation에 전달하는 callLLM 콜백도 동일하게 폴백 적용
- 응답에 fallbackChain/fallbackUsed 추가되지만 호출자는 무시 가능

테스트: 전체 2635 통과 (회귀 0). mock된 callLLM이 fallback 모듈을 통해 호출되어
기존 SDK 테스트가 그대로 작동.
