+++ src/executor.js
@@ -27,7 +27,9 @@ import {
   orchestrateConsultation,
   enrichTaskOutputWithConsultation,
 } from '../scripts/lib/engine/expert-consultation.js';
-import { callLLM } from '../scripts/lib/llm/llm-provider.js';
+// callLLMWithFallback을 callLLM 별칭으로 사용 — 첫 모델이 429/5xx로 실패하면
+// ModelSelector.selectFallback 체인 따라 자동 다음 모델 시도. 시그니처/응답 호환.
+import { callLLMWithFallback as callLLM } from '../scripts/lib/llm/llm-fallback.js';
 import {
   createPhaseWorktree,
   removePhaseWorktree,
